refactor: move scenario_01-specific steps from common_steps to scenario_01
